+++ tket/src/Transformations/Transform.hpp
@@ -453,6 +453,14 @@ class Transform {
   // Produces: Any gates
   static Transform commute_through_multis();
 
+  // replaces every non-global PhasedX gate with two global PhasedX gates,
+  // using:
+  //
+  //         PhX(α, β) = PhX(-1/2, β + 1/2) Rz(α) PhX(1/2, β + 1/2)
+  //
+  // Setting α = 0 on non-targeted qubits makes the RHS the identity
+  static Transform globalise_phasedx();
+
   // commutes Rz gates through ZZMax, and combines adjacent ZZMax gates
   // Expects: ZZMax, Rz, Rx
   // Produces: ZZMax, Rz, Rx